Refactor and switch to django
