+++ chat/urls.py
@@ -5,8 +5,8 @@ urlpatterns = [
     path('', views.index, name='index'),
     path('about/', views.about, name='about'),
     path('chat/', views.chat, name='chat'),
-    path('chat/response/', views.get_chat_response, name='get_chat_response'),
-    path('chat/upload/', views.upload_file, name='upload_file'),
-    path('api/chat_history/', views.get_chat_history, name='get_chat_history'),
-    path('api/clear_history/', views.clear_history, name='clear_history'),
+    path('api/chat/response/', views.get_chat_response, name='get_chat_response'),
+    path('api/chat/upload/', views.upload_file, name='upload_file'),
+    path('api/chat/history/', views.get_chat_history, name='get_chat_history'),
+    path('api/chat/clear_history/', views.clear_history, name='clear_history'),
 ]
