+++ chat/utils/__init__.py
@@ -1,0 +1,3 @@
+from .markdown import render_html_response
+
+__all__ = ['render_html_response']
