+++ templates/chat/bot_message.html
