+++ templates/chat/chat.html
@@ -3,8 +3,9 @@
 {% block title %}Analysis | EDA Chatbot{% endblock %}
 
 {% block content %}
-<div class="container mt-4">
-    <form class="mb-4" hx-post="/upload" hx-target="#chat-container" hx-swap="beforeend" enctype="multipart/form-data" id="upload-form">
+<div class="container">
+    <form class="mb-3" hx-post="/api/chat/upload/" hx-target="#chat-container" hx-swap="beforeend" enctype="multipart/form-data" id="upload-form">
+        {% csrf_token %}
         <div class="input-group">
             <input type="file" name="file" class="form-control" accept=".csv" title="Max file size: 2 MB. Only CSV files are allowed." required>
             <button class="btn btn-secondary" type="submit">Upload Dataset</button>
@@ -17,7 +18,8 @@
         </div>
         <!-- Chat messages will be inserted here -->
     </div>
-    <form class="mt-3" hx-post="/chat" hx-target="#chat-container" hx-swap="beforeend" hx-indicator="#typing-indicator" id="message-form">
+    <form class="mt-3" hx-post="/api/chat/response/" hx-target="#chat-container" hx-swap="beforeend" hx-indicator="#typing-indicator" id="message-form">
+        {% csrf_token %}
         <div class="input-group">
             <input type="text" name="message" id="message-input" class="form-control" placeholder="Type your message..." autocomplete="off" required>
             <button class="btn btn-primary" type="submit">Send</button>
