+++ templates/pages/about.html
@@ -128,12 +128,12 @@
                     <div class="text-center">
                         <div class="video-container">
                             <video 
-                                src="/static/demo-vid.mp4" 
+                                src="/static/assets/demo-vid.mp4" 
                                 class="img-fluid rounded" 
                                 style="max-width: 80%; height: auto;" 
                                 controls
                                 preload="metadata"
-                                poster="/static/demo-loading-placeholder.png"
+                                poster="/static/assets/demo-loading-placeholder.png"
                             ></video>
                             <div class="video-loading">
                                 <div class="spinner-border text-primary" role="status">
