+++ templates/pages/index.html
@@ -18,7 +18,7 @@
 
     <!-- Demo Image Section -->
     <div class="text-center mt-4">
-        <img src="static/demo.png" alt="Demo of the App" class="img-fluid rounded shadow" style="max-width: 80%; height: auto;">
+        <img src="static/assets/demo.png" alt="Demo of the App" class="img-fluid rounded shadow" style="max-width: 80%; height: auto;">
         <p class="mt-3 text-muted">A quick preview of the Exploratory Data Analysis App in action.</p>
     </div>
